% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_week.R
\name{get_start_week}
\alias{get_start_week}
\title{Test function}
\usage{
get_start_week(x)
}
\arguments{
\item{x}{Get the startdate of that week (Monday).}
}
\value{

}
\description{
Toy example to init empty package structure. Returns the date on which the week started.
}
\examples{
get_start_week(Sys.Date())
}
